pyproject.toml
[tool.poetry]
name = "aas-middleware"
version = "0.2.3"
description = ""
authors = ["Sebastian Behrendt <sebastianbehrendt97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.1"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
basyx-python-sdk = "^1.0.0"
graphene-pydantic = "^0.6.1"
anyio = "^4.3.0"
aiohttp = "^3.10.11"
toml = "^0.10.2"
ba-syx-aas-environment-component-client = "^2.1"
datamodel-code-generator = "^0.25.6"
importlib-metadata = "^8.0.0"

aiomqtt = { version = "^2.1.0", optional = true }
asyncua = { version = "^1.1.0", optional = true }
mkdocstrings = {extras = ["python-legacy"], version = "^0.25.1"}
starlette-graphene3 = "^0.6.0"
typeguard = "^4.4.1"
exceptiongroup = "^1.2.2"
aas-pydantic = "^0.1.0"

[tool.poetry.extras]
industrial = ["aiomqtt", "asyncua"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"
pytest-order = "^1.2.1"
mkdocs-material = "^9.5.29"
mkdocs-include-dir-to-nav = "^1.2.0"
snakeviz = "^2.2.0"

[tool.poetry.group.sdm.dependencies]
pandas = "^2.2.2"
prodsys = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
